requirements.txt
pandas >= 2.1.0
numpy >= 1.24.0
scikit-learn >= 1.3.0
scipy >= 1.11.0
matplotlib>=3.7.0
seaborn >= 0.12.0
pyproj >= 3.6.0
geopy >= 2.4.0
Pillow >= 10.0.0
opencv-python >= 4.8.0
shapely >= 2.0.0
rasterio >= 1.3.0
requests >= 2.31.0
SQLAlchemy >= 2.0.0
shap >= 0.42.1